AsRef<Path> -> Into<PathBuf>
Previously, this code said "Hey, I will take anything that can be
converted into a &Path," and now it says "Hey I can take anything that
can be turned into a PathBuf."

The former is not *wrong*, but is a bit roundabout. You want to store a
PathBuf. Just take anything that can be turned into a PathBuf. Not only
is this a little simpler, but it also can save you an allocation, at
least conceputally, I don't know if rustc is smart enough to remove the
allocation in code like this:

```rust
let path = PathBuf::from("/etc/passwd");
let manager = RegistryManager::new(&path);
```

With the previous version of the code, this would convert a PathBuf into
a &Path, by using a pointer to it, and then creates a whole new PathBuf
out of it, by allocating. This way instead, we can still take the same
stuff, but if we have a PathBuf already, we skip the roundtrip through a
pointer. In other words, you could just pass `path` in the code above,
whereas you couldn't with the original version.