Remove an unwrap
You're already returning Result, best to not panic.

I decided to try and keep this in the same style you currently have,
with a String living in the error type, but you could also just stash
the io::Error in there too.

Also, I didn't make this change in this diff beacuse it is up to you,
but I would not be making a new module for the erorr, I would put it in
lib.rs. You're not really buying anything by moving it in there, just
making yourself import it when you don't have to. Up to you though!