or -> or_else
So, because we do some work in the conversion of these error messages,
*technically* this way is slightly more efficient. You see, with or, we
unconditionally also run the "Generate the error" code. But with
or_else, we only invoke that code if we actually hit the error here. So
in effect, doing this moves the allocation to the error path only, which
is nice.